requirements.txt
tensorflow
Pillow
matplotlib
pandas
tflearn
sklearn
scipy


